cleanup actions.py, loggers
