+++ squidwork/logger/__init__.py
@@ -1,3 +1,2 @@
-from .logger import Logger
 from .logger_to_logfile import LoggerToLogfile
 from .logger_to_stdout import LoggerToStdout
